Update .clabot
I propose adding the following line to the cla-bot-message and removing the @ mentions:
"Once you have successfully signed the FINOS CLA you must email help@finos.org. Please include your full name, email address, and links to relevant PRs."

The reason for this is that contributors need to be added manually. If they know to email us it makes it more manageable.